+++ gradle-test-extensions-subprojects/gradle-testkit-test-extensions/main/kotlin/com/javiersc/gradle/testkit/test/extensions/_internal/argumentsTxt.kt
@@ -1,32 +1,6 @@
-@file:OptIn(ExperimentalStdlibApi::class)
+package com.javiersc.gradle.testkit.test.extensions._internal
 
-package com.javiersc.gradle.testkit.test.extensions
-
-import java.io.File
-import java.nio.file.Files.createTempDirectory
-import java.nio.file.Path
-import java.nio.file.Paths
 import org.gradle.testkit.runner.GradleRunner
-
-public fun resourceFile(resource: String): File =
-    File(Thread.currentThread().contextClassLoader?.getResource(resource)?.toURI()!!)
-
-internal val sandboxPath: Path = Paths.get("build/sandbox").apply { toFile().mkdirs() }
-
-internal val sandboxIsolatedPath: Path =
-    Paths.get("build/sandbox-isolated").apply { toFile().mkdirs() }
-
-internal fun String.copyResourceTo(destination: File) {
-    resourceFile(this).copyRecursively(destination)
-}
-
-internal fun createSandboxDirectory(prefix: String): File =
-    createTempDirectory(sandboxPath, if (prefix.isBlank()) prefix else "$prefix-").toFile()
-
-internal fun createSandboxIsolatedDirectory(prefix: String): File =
-    createTempDirectory(sandboxIsolatedPath, if (prefix.isBlank()) prefix else "$prefix-")
-        .toFile()
-        .absoluteFile
 
 internal val GradleRunner.argumentsTxt: List<String>
     get() {